+++ cg/apps/lims/limsjson.py
@@ -28,6 +28,8 @@ def get_project_type(samples: [dict]) -> str:
         project_type = "mip-dna"
     elif data_analyses == {"fluffy"}:
         project_type = "rml"
+    elif data_analyses == {"balsamic"}:
+        project_type = "balsamic"
     else:
         raise OrderFormError(f"Unsupported json orderform: {data_analyses}")
 
@@ -80,7 +82,7 @@ def expand_case(case_id: str, parsed_case: dict) -> dict:
 
     gene_panels = set()
     for raw_sample in samples:
-        if raw_sample["panels"]:
+        if raw_sample.get("panels"):
             gene_panels.update(raw_sample["panels"])
 
         new_sample = {}
@@ -96,6 +98,7 @@ def expand_case(case_id: str, parsed_case: dict) -> dict:
 
         new_case["samples"].append(new_sample)
 
-    new_case["panels"] = list(gene_panels)
+    if gene_panels:
+        new_case["panels"] = list(gene_panels)
 
     return new_case