+++ CHANGELOG.md
@@ -11,8 +11,14 @@ Try to use the following format:
 ## [x.x.x]
 
 ### Added
-
+### Changed
 ### Fixed
+
+## [NG.NG.NG]
+
+### Added
+
+- Added support for Balsamic-orderforms as json-file
 
 ### Changed